created some components and managed to app to app
